requirements.txt
argparse==1.1
PyYAML==6.0.3
opencv-python>=4.11.0
numpy>=2.2.4
